requirements.txt
pandas==2.2.0
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.1.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
